requirements.txt
Flask
python-dotenv
requests
scikit-learn
